:building_construction: move lifecycle.js
